doing +43

+++ OneLib/stdcxx/LiteMutableBuffer.cc
@@ -1,6 +1,6 @@
 #include "stdcxx/LiteMutableBuffer.h"
 
-bool LiteMutableBuffer::initWithCapacity(size_t in_capacity) {
+bool LiteMutableBuffer::initWithCapacity(int in_capacity) {
   buffer_ = (byte *)LiteMemOpt::alloc(in_capacity);
   if (!buffer_) {
     return false;
@@ -11,13 +11,13 @@ bool LiteMutableBuffer::initWithCapacity(size_t in_capacity) {
   return true;
 }
 
-size_t LiteMutableBuffer::ensureCapacity(size_t new_capacity) {
+int LiteMutableBuffer::ensureCapacity(int new_capacity) {
   byte *new_buffer;
   unsigned int final_capacity;
 
   if (new_capacity <= capacity_)
     return capacity_;
-  final_capacity = ALIGN(new_capacity, 8);
+  final_capacity = (int)ALIGN(new_capacity, 8);
 
   new_buffer = (byte *)LiteMemOpt::alloc(final_capacity);
